migrated common code of ipaddress_tests.py to a file called ipaddress_utils.py
